% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-data-2-meta.R
\name{add_data_parties}
\alias{add_data_parties}
\title{Extract additional data from text and meta data}
\usage{
add_data_parties(data_case)
}
\description{
Various functions for importing extract additional data from text
and meta data
}

