% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domstolr-import.R
\name{domstolr_import}
\alias{domstolr_import}
\title{Import selections downloaded as html from Lovdata Pro (C)}
\usage{
domstolr_import(file = NULL, directory = NULL, regex = ".*.html$",
  recursive = TRUE, meta_only = FALSE, match_judges = TRUE,
  verbose = FALSE)
}
\arguments{
\item{file}{Single file to import.}

\item{directory}{Directory to import files from.}

\item{regex}{Regular Expression to use when searching for files within directory. Default is to extract all html files.}

\item{recursive}{If TRUE then it will also search for files within subfolders of the provided directory.}

\item{meta_only}{If TRUE it will only return the data within the header table of each case.}

\item{verbose}{If TRUE it will print out dots for each parsed case to signal how far the function have come.}
}
\description{
Import selections downloaded as html from Lovdata Pro (C)
}
\examples{
\dontrun{
domstol_data <- domstolr_import(directiory = "data/")
save(data, file = "domstol_data.RData")
}
}
\keyword{domstolr}
\keyword{lovdata}

