% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-data-1-html.R
\name{extract_data_html}
\alias{extract_data_html}
\title{Extract data from the html code}
\usage{
extract_data_html(.case, data_meta, all_tables, verbose)
}
\description{
Takes the html code from one case and extracts text and
references. The html must be parsed differently depending on when
the case is from.
}

