% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_cases}
\alias{get_data}
\alias{get_judges}
\alias{get_keywords}
\alias{get_paragraphs}
\alias{get_parties}
\alias{get_proceedings}
\alias{get_references}
\title{Extract data as data frame from domstolr data}
\usage{
get_data(data, type = "cases")

get_cases(data)

get_paragraphs(data)

get_references(data)

get_parties(data)

get_proceedings(data)

get_judges(data)

get_keywords(data)
}
\arguments{
\item{data}{A domstolr object (from \code{domstolr_import()}).}

\item{type}{Which data frame to extract (cases, paragraphs,
references, parties, proceedings, judges, or keywords).}
}
\description{
Extract data as data frame from domstolr data
}
\examples{
\dontrun{
domstol_data <- domstolr_import(directiory = "data/")
cases <- get_cases(domstol_data)
}
}
\keyword{cases}
\keyword{domstolr}
\keyword{extract}
\keyword{judges}
\keyword{keywords}
\keyword{paragraps}
\keyword{parties}
\keyword{proceedings}
\keyword{references}

